core: Move select.rs to translate/
