+++ core/translate/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod expr;
+pub(crate) mod select;
 pub(crate) mod where_clause;
 
 use std::cell::RefCell;
@@ -8,7 +9,7 @@ use expr::{build_select, maybe_apply_affinity, translate_expr};
 use crate::function::{AggFunc, Func};
 use crate::pager::Pager;
 use crate::schema::{Column, PseudoTable, Schema, Table};
-use crate::select::{ColumnInfo, LoopInfo, Select, SrcTable};
+use crate::translate::select::{ColumnInfo, LoopInfo, Select, SrcTable};
 use crate::sqlite3_ondisk::{DatabaseHeader, MIN_PAGE_CACHE_SIZE};
 use crate::types::{OwnedRecord, OwnedValue};
 use crate::util::normalize_ident;
