+++ core/translate/where_clause.rs
@@ -4,7 +4,7 @@ use sqlite3_parser::ast::{self, JoinOperator};
 use crate::{
     translate::expr::{resolve_ident_qualified, resolve_ident_table, translate_expr},
     function::SingleRowFunc,
-    select::Select,
+    translate::select::Select,
     vdbe::{BranchOffset, Insn, ProgramBuilder},
 };
